a2a_agent/requirements.txt
a2a-sdk
google-adk
google-genai
google-cloud-aiplatform
google-cloud-bigquery
pandas
numpy
matplotlib
seaborn
immutabledict
regex
sqlglot
praw
uvicorn
starlette
gunicorn
python-dotenv
asyncclick
requests